Changing order so that context handler runs before resource handler.
The resource handler should only run after no context handlers handles the URL. If a servlet maps to /foo.js and the resource handler responds to foo.js, it's the servlet that should handle the request. This also fixes servlets that maps to "/" that the resource handler will claim to have "handled" regardless of whether or not an index.html exists.