Correctly handling priority between servlet context mappings and static resources.
Turns out that the initial implementation where the static resources
took presedence over mapped context was wrong. The mapped contexts
always gets presedence in tomcat, jetty, etc.

This means that we can piggyback on Jetty's built in ResourceCollection,
since all we really want is for Jetty to look for our resources in both
the build directory and in the src/main/webapp directory that the war
plugin uses.